Adding multiple headers

+++ logstore/src/test/java/io/joshworks/fstore/log/appender/compaction/combiner/UniqueMergeCombinerTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertEquals;
 
 public class UniqueMergeCombinerTest {
 
-    private static final long MAX_ENTRY_SIZE = 1024 * 1024 * 5L;
+    private static final int MAX_ENTRY_SIZE = 1024 * 1024 * 5;
     private static final double CHCKSUM_PROB = 1;
 
     private final List<Segment> segments = new ArrayList<>();
